Update Switch-Case-Blackjack.js
took the "currPlayerHand" conditional ternary operator out of the for loop as it was causing issues when the player drew cards that totalled less than 11. I also changed the condition to check from 11 (don't know why I chose that?) to less than 4 which is the actual minimum a player can draw with 2 cards. It now works.